+++ src/main/java/de/dm/intellij/liferay/util/Icons.java
@@ -9,5 +9,6 @@ public class Icons {
     public static final Icon BND_ICON = IconLoader.getIcon("/org/bndtools/bnd/bnd.png", Icons.class);
     public static final Icon LIFERAY_ICON = IconLoader.getIcon("/com/liferay/icons/liferay.png", Icons.class);
     public static final Icon FREEMARKER_ICON = IconLoader.getIcon("/com/liferay/icons/freemarker.png", Icons.class);
+    public static final Icon VELOCITY_ICON = IconLoader.getIcon("/com/liferay/icons/velocity.svg", Icons.class);
 
 }